+++ sdk/src/androidMain/kotlin/com/iproov/kmp/api/Iproov.android.kt
@@ -7,6 +7,7 @@ import androidx.core.content.edit
 import com.iproov.kmp.AppContext
 import com.iproov.kmp.mapper.toIproov
 import com.iproov.kmp.mapper.toIproovState
+import com.iproov.kmp.mapper.toIproovUIState
 import com.iproov.sdk.api.IProov
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -22,11 +23,13 @@ actual object Iproov {
     private val job = SupervisorJob()
     private val coroutineScope = CoroutineScope(Dispatchers.IO + job)
     private var sessionStateJob: Job? = null
+    private var sessionUIStateJob: Job? = null
 
     actual val sdkVersion: String = IProov.sdkVersion
     actual val buildNumber: String = IProov.buildNumber
 
     actual val sessionState: MutableStateFlow<IproovState?> = MutableStateFlow(null)
+    actual val uiState: MutableStateFlow<IproovUIState?> = MutableStateFlow(null)
 
     init {
         setEnvironment(AppContext.get())
@@ -35,15 +38,28 @@ actual object Iproov {
     actual suspend fun launchSession(baseUrl: String, token: String, iproovOptions: IproovOptions) {
         val session = IProov.createSession(AppContext.get(), baseUrl, token, iproovOptions.toIproov())
         // Observe first, then start
-        observeSessionState(session) {
-            session.start()
-        }
+        observeSessionState(session) { session.start() }
+        observeSessionUIState(session)
     }
 
     actual suspend fun cancelSession() {
         IProov.session?.cancel()
         sessionStateJob?.cancel()
         sessionState.value = null
+    }
+
+    private fun observeSessionUIState(session: IProov.Session) {
+        sessionUIStateJob?.cancel()
+        sessionUIStateJob = coroutineScope.launch(Dispatchers.IO) {
+            session.uiState
+                .collect { sessionUIState ->
+                    if (sessionUIStateJob?.isActive == true) {
+                        withContext(Dispatchers.Main) {
+                            uiState.value = sessionUIState.toIproovUIState()
+                        }
+                    }
+                }
+        }
     }
 
     private fun observeSessionState(session: IProov.Session, whenReady: (() -> Unit)? = null) {