+++ sdk/src/androidMain/kotlin/com/iproov/kmp/mapper/IproovMapper.kt
@@ -3,22 +3,25 @@
 package com.iproov.kmp.mapper
 
 import android.graphics.Bitmap
-import com.iproov.sdk.api.IProov
+import com.iproov.kmp.AppContext
 import com.iproov.kmp.api.CameraException
 import com.iproov.kmp.api.CameraPermissionException
 import com.iproov.kmp.api.Canceler
 import com.iproov.kmp.api.CaptureAlreadyActiveException
-import com.iproov.kmp.api.FaceDetectorException
+import com.iproov.kmp.api.FailureReason
 import com.iproov.kmp.api.FailureResult
 import com.iproov.kmp.api.IProovException
 import com.iproov.kmp.api.InvalidOptionsException
 import com.iproov.kmp.api.IproovState
+import com.iproov.kmp.api.IproovUIState
 import com.iproov.kmp.api.MultiWindowUnsupportedException
 import com.iproov.kmp.api.NetworkException
 import com.iproov.kmp.api.ServerException
 import com.iproov.kmp.api.SuccessResult
 import com.iproov.kmp.api.UnexpectedErrorException
 import com.iproov.kmp.api.UnsupportedDeviceException
+import com.iproov.sdk.api.IProov
+import com.iproov.sdk.api.IProov.UIState
 import java.io.ByteArrayOutputStream
 
 
@@ -35,9 +38,10 @@ fun IProov.State.toIproovState(): IproovState {
 
         is IProov.State.Failure -> {
             val frame = if (this.failureResult.frame == null) null else bitmapToByteArray(this.failureResult.frame!!)
-            val reason = this.failureResult.reason
+            val context = AppContext.get()
+            val reasons = this.failureResult.reasons.map { FailureReason(it.feedbackCode, context.getString(it.description)) }
 
-            IproovState.Failure(FailureResult(reason.name, reason.feedbackCode, frame))
+            IproovState.Failure(FailureResult(reasons, frame))
         }
 
         is IProov.State.Canceled -> {
@@ -52,6 +56,15 @@ fun IProov.State.toIproovState(): IproovState {
     }
 }
 
+fun UIState.toIproovUIState(): IproovUIState {
+    return when (this) {
+        UIState.NotStarted -> IproovUIState.NotStarted
+        UIState.Started -> IproovUIState.Started
+        UIState.Ended -> IproovUIState.Ended
+    }
+}
+
+
 private fun com.iproov.sdk.api.exception.IProovException.toIproovException(): IProovException {
     return when (this) {
         is com.iproov.sdk.api.exception.CaptureAlreadyActiveException -> CaptureAlreadyActiveException(this.reason)
@@ -60,7 +73,6 @@ private fun com.iproov.sdk.api.exception.IProovException.toIproovException(): IP
         is com.iproov.sdk.api.exception.ServerException -> ServerException(this.reason, this.message)
         is com.iproov.sdk.api.exception.CameraException -> CameraException(this.reason, this.message)
         is com.iproov.sdk.api.exception.MultiWindowUnsupportedException -> MultiWindowUnsupportedException(this.reason)
-        is com.iproov.sdk.api.exception.FaceDetectorException -> FaceDetectorException(this.reason, this.message)
         is com.iproov.sdk.api.exception.UnsupportedDeviceException -> UnsupportedDeviceException(this.reason)
         is com.iproov.sdk.api.exception.InvalidOptionsException -> InvalidOptionsException(this.reason)
         else -> UnexpectedErrorException(this.reason, this.message)