+++ sdk/src/iosMain/kotlin/com/iproov/kmp/api/Iproov.ios.kt
@@ -7,6 +7,7 @@ import com.iproov.kmp.mapper.toIproov
 import com.iproov.sdk.IPErrorCode
 import com.iproov.sdk.IPSession
 import com.iproov.sdk.IProov
+import com.iproov.sdk.IProovViewDelegateProtocol
 import com.iproov.sdk.buildNumber
 import com.iproov.sdk.launchWithStreamingURL
 import com.iproov.sdk.versionStr
@@ -14,6 +15,7 @@ import kotlinx.cinterop.ExperimentalForeignApi
 import kotlinx.coroutines.flow.MutableStateFlow
 import platform.Foundation.NSURL
 import platform.Foundation.NSUserDefaults
+import platform.darwin.NSObject
 
 
 @OptIn(ExperimentalForeignApi::class)
@@ -23,6 +25,7 @@ actual object Iproov {
     actual val buildNumber: String = IProov.buildNumber().toString()
 
     actual val sessionState: MutableStateFlow<IproovState?> = MutableStateFlow(null)
+    actual val uiState: MutableStateFlow<IproovUIState?> = MutableStateFlow(null)
 
     private var session: IPSession? = null
 
@@ -33,6 +36,8 @@ actual object Iproov {
     actual suspend fun launchSession(baseUrl: String, token: String, iproovOptions: IproovOptions) {
         val url = NSURL(string = baseUrl)
         val options = iproovOptions.toIproov()
+
+        options.setViewDelegate(delegate)
 
         session = IProov.launchWithStreamingURL(
             streamingURL = url,
@@ -54,12 +59,15 @@ actual object Iproov {
             canceled = {
                 sessionState.tryEmit(IproovState.Canceled(Canceler.USER))
             },
-            failure = {
-                val frame = if (it?.frame() == null) null else convertUIImageToByteArray(it.frame()!!)
-                val reason = it?.reason()
-                val feedbackCode = it?.localizedDescription() ?: ""
+            failure = { feedbackCodeList, descriptionList, frame ->
+                val frameReceived = if (frame == null) null else convertUIImageToByteArray(frame)
 
-                sessionState.tryEmit(IproovState.Failure(FailureResult(reason.toString(), feedbackCode, frame)))
+                val failureReasons = mutableListOf<FailureReason>()
+                feedbackCodeList?.forEachIndexed { index, feedbackCode ->
+                    failureReasons.add(FailureReason(feedbackCode.toString(), descriptionList?.get(index)?.toString() ?: ""))
+                }
+
+                sessionState.tryEmit(IproovState.Failure(FailureResult(failureReasons, frameReceived)))
             },
             error = {
                 val description = it?.localizedDescription() ?: ""
@@ -78,8 +86,23 @@ actual object Iproov {
         )
     }
 
+
     actual suspend fun cancelSession() {
         session?.cancel()
+    }
+
+    private val delegate = object : NSObject(), IProovViewDelegateProtocol {
+        override fun willPresentIProovView() {
+            uiState.value = IproovUIState.NotStarted
+        }
+
+        override fun didPresentIProovView() {
+            uiState.value = IproovUIState.Started
+        }
+
+        override fun didDismissIProovView() {
+            uiState.value = IproovUIState.Ended
+        }
     }
 
     private fun setEnvironment() {