+++ CHANGELOG.md
@@ -1,17 +1,37 @@
-# iProov Biometrics Kotlin-Multiplatform SDK
+# iProov Face Kotlin-Multiplatform SDK
+
+## 2.0.0
+
+iProov Face Kotlin-Multiplatform SDK v2.0.0 includes the following:
+
+- Adding the new Options in GenuinePresenceAssurance(controlYPosition, controlXPosition, and scanningPrompts)
+- New UI state API
+
+### iOS
+
+* SDK version [v13.1.0](https://github.com/iProov/ios/releases/tag/13.1.0).
+
+### Android
+
+* SDK version [v11.1.0](https://github.com/iProov/android/releases/tag/v11.1.0).
+
 
 ## 1.0.1
 
-iProov Biometrics Kotlin-Multiplatform SDK v1.0.1 includes the following:
+iProov Face Kotlin-Multiplatform SDK v1.0.1 includes the following:
 
 ### iOS
 
 * SDK version [v12.4.1](https://github.com/iProov/ios/releases/tag/12.4.1).
 
+### Android
+
+* SDK version [v10.3.2](https://github.com/iProov/android/releases/tag/v10.3.2).
+
 
 ## 1.0.0
 
-iProov Biometrics Kotlin-Multiplatform SDK v1.0.0 includes the following:
+iProov Face Kotlin-Multiplatform SDK v1.0.0 includes the following:
 
 ### iOS